get_field_string not working if it must allocate
The get_field_string_* functions allocates memory if the passed
destination's size is set to 0. However, this does not work and
generates a memory leak. Because there is not yet a case where the
functions must allocate memory, the issue does not yet happen.

To allow to let the function allocate the required space, functions are
changed. The get_field_string_*() don't allocate themselves, and simply
returns the pointer to the found string.

Macros are adapted to remain compatible with current usage.

Signed-off-by: Stefano Babic <sbabic@denx.de>