Add examples to build systems
Some of the tools in examples/ are utilities and they could be installed
into the target. Add these utilities to the build system to make easier
to integrate them. Move all these utilities in a  more general "tools"
directory and add them to the build system.

Reorder Makefile to reuse the same command for all binaries.

Signed-off-by: Stefano Babic <sbabic@denx.de>