add libressl support
Because LibreSSL returns 2.x for the version numbers, the macros think
that OpenSSL 1.1 should be used. This patch checks to see if there is
an LIBRESSL_VERSION_NUMBER defined.

Also, add a check to see if CONFIG_SIGALG_CMS is enabled in sslapi.h
before including openssl/cms.h.  LibreSSL does not support cms and
probably never will.  As such, LibreSSL doesn't include it and will
fail during a build that would otherwise pass.

Signed-off-by: Adam Duskett <Adamduskett@outlook.com>