hawkbit: artifact not installed if SWU is not the first
If the Hawkbit server delivers artifacts not in SWU format, SWUpdate
stops. The behaviour is changed and SWUpdate will simply skip all
artifacts that are not SWU images. This avoids that Software is not
installed at all if SWU is not the first artifact in the list provided
by the server.

Signed-off-by: Stefano Babic <sbabic@denx.de>