core: don't unlink if final equals temp location
If the temporary location of an artifact happens to
be identical to the final location of the artifact,
SWUpdate deletes the artifact after installation
(if CONFIG_NOCLEANUP isn't set, that is).

Consider the following sw-description with default
TMPDIR=/tmp set:

software =
{
    version = "0.0.1";
    files: (
        {
            filename = "dummy.file";
            path = "/tmp/dummy.file";
        }
    );
}

After installation, /tmp/dummy.file will be removed.
Prevent its removal and issue a warning. Additionally,
if dummy.file is referenced more than once, issue a
warning stating its removal reason.

Signed-off-by: Christian Storm <christian.storm@siemens.com>