logging: forward loglevel to systemd's journald
If SWUpdate is run as a systemd service, its stdout and stderr
output is per default sent to journald. journald supports
specifying the log priority by prefixing the log messages with
priority values enclosed in < >, following the scheme used by
the kernel's printk().

Adapt the console_notifier to emit priority-prefixed log messages
if SWUpdate is run as systemd service.

Signed-off-by: Christian Storm <christian.storm@siemens.com>
Acked-by: Stefano Babic <sbabic@denx.de>