core: do install artifact if final or temp locations are empty
Commit d7b7da4 "core: don't unlink if final equals temp location"
prevents artifact deletion if the temporary location of the artifact
happens to be identical to the final location.

However, img->path and img->extract_file are not always non-empty,
e.g., for an artifact of type "bootloader" installed from a local
source installed via swupdate -i:

software =
{
    version = "0.0.1";
    images: (
        {
            filename = "bootloader.cfg";
            type = "bootloader";

        });
}

Hence, add a check for this condition.

Signed-off-by: Christian Storm <christian.storm@siemens.com>
Acked-by: Stefano Babic <sbabic@denx.de>