Lua: make handler not found message more pleasant
Currently, a stack dump is printed on a failed
require("swupdate_handlers") which is thereafter
clarified as being no error.
Make this message more friendly.

Only package.path is shown as it's assumed that
custom Lua handlers won't be written as a .so.
In this case, they may be written as a regular
handler in C in the first place.

Signed-off-by: Christian Storm <christian.storm@siemens.com>