Lua: option to embed Lua handler into SWUpdate binary
Lua handlers are loaded at run-time, i.e., a
    require("swupdate_handlers")
is executed at startup. Hence, the Lua handler source
code file swupdate_handlers.lua has to be deployed on
the target system respecting Lua's search path.

The new config option EMBEDDED_LUA_HANDLER allows to
embed the Lua handler's source into the SWUpdate binary.
Hence, (1) it is independent of the Lua search path on
the target system and (2) it is harder to tamper with.

Signed-off-by: Christian Storm <christian.storm@siemens.com>