Lua: don't extract image prior to handler call
In order to be able to call copyfile() from within a Lua handler
or when chaining handlers, i.e., calling a C handler from within
a Lua handler, the image shouldn't be extracted beforehand as
this extraction, e.g., moves img_type's offset so that the
Lua handler would have to seek back prior to calling, e.g.,
copyfile().

Instead of extracting prior to calling the Lua handler, rely on
the Lua handler either to do or to instruct the image's extraction,
e.g., via calling copyfile().

NOTE: This is a change in behavior and breaks current Lua handler
deployments as they expect the image to be extracted.

Signed-off-by: Christian Storm <christian.storm@siemens.com>