Lua: introduce method to chain-call C handlers
As the artifact's data is not unconditionally extracted and
img_type's fdin is available to Lua, a Lua handler may chain-call
a C handler to process the artifact while having the ability to
do work in Lua prior to and/or after having called the C handler.

Available C handlers are exposed to Lua in the swupdate.handlers
table (as keys). The new Lua function swupdate.call_handler() can
then be used to chain-call a C handler by providing the
C handler's name and the image parameter to it.

Signed-off-by: Christian Storm <christian.storm@siemens.com>