Lua: introduce image:read() method
For C handlers, there's core/cpio_util.c's copyfile() as
a general method to extract data from the input stream and
copy it to somewhere. image:read(<callback()>) is the
analogous function exposed to Lua that reads from the input
stream and calls the callback Lua function <callback()> for
every chunk read, passing this chunk as parameter.

Signed-off-by: Christian Storm <christian.storm@siemens.com>