Add a first-in, first-install order
Currently, files are installed before images and each entry
in sw-description is installed in reverse order. In some applications,
the installation relies on the order, for example if files are installed on a fresh
burned image.

To provide this use case, install "images" before "files" and ensures that they
are installed in the same order they are put into the sw-description file.

Signed-off-by: Stefano Babic <sbabic@denx.de>