handler: swupdate forwarder
Add a handler to forward SWU images. A global SWU can contain
other SWU that are forwarded to other devices (or subsystems)
running SWUpdate as well.

The handler transfers an embedded SWU to all destinations assigned
into sw-description at the same time (streaming works with the handler).

The destinations are defined in properties for the entry:

{
	filename = "image.swu";
	type = "swuforward";

	properties: ({
		name = "url";
		value = "http://<dest ip>:8080";
		}
	);
}

The handler assumes that the destinations are running SWUpdate
with activated Webserver and uses the REST-API to send the image.

Signed-off-by: Stefano Babic <sbabic@denx.de>