flash_handler: skip writing to 0xff filled pages on nand
This is necessary when flashing an ubi image. A block filled with 0ff
does not have an ECC made of 0xff on certein hardware ECC alorithms, So
this already flips some bits, when written.
However, ubi treats a page full of 0xff as empty, which is
not true here. And the result is that data written to such an empty
page yields ECC Errors.

This is basically Code taken from nand-utils/nandwrite.c from
git://git.infradead.org/mtd-utils.git

See http://lists.infradead.org/pipermail/linux-mtd/2016-December/070793.html
for the patch that starts the discussion and the mtd ML.

The relevant Commits in mtd-utils are the following ones:

---------------------------------------------------------------------------
commit 15c21334b201dc54870cfd3e9697307c95f7e4dc
Author: Kees Trommel <ctrommel@linvm302.aimsys.nl>
Date:   Tue Dec 6 09:21:19 2016 +0100

    nandwrite: add skip-all-ff-pages-option

    Signed-off-by: Kees Trommel <ctrommel@linvm302.aimsys.nl>
    Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>

commit f5b47ef5112326b06fd875982379769c55a71703
Author: Marek Vasut <marex@denx.de>
Date:   Fri Dec 9 02:47:41 2016 +0100

    nandwrite: Factor out buffer checking code

    Pull the buffer content checking code into separate function and
    simplify the code invoking it slightly.

    Signed-off-by: Marek Vasut <marex@denx.de>
    Reviewed-by: Boris Brezillon <boris.brezillon@free-electrons.com>
    Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>
---------------------------------------------------------------------------

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Acked-by: Stefano Babic <sbabic@denx.de>