Do not skip sectors to erase if unlock is not supported
Not all flashes support lock, and mtd_is_locked() return true
even if unlock is unsupported, and the sector is not erased.
This generates an error later when sector is written.

Check the error code and ignore it if the feature is unsupported
for the MTD device to be erased.

Signed-off-by: Stefano Babic <sbabic@denx.de>