LUA: export SWUpdate function to LUA scripts
SWUpdate function are exported for LUA handlers, but not for LUA
scripts. A run_lua_script() function was prepared, but never used
and the LUA Script Handler duplicates most of code.

Factorize the LUA Script Handler to use run_lua_script() and exports
SWUpdate's function for scripts. This allows to use
	require("swupdate")

in own LUA scripts.

Signed-off-by: Stefano Babic <sbabic@denx.de>