Encrypted scripts cannot be executed
Scripts cannot be executed because they are not decrypted
before they run. The bug happens when the SWU is downloaded,
while a local install is working because scripts are extracted
and decrypted via copyfile().

Factorize script extraction between local and remote install
and always decrypt scripts before passing them to the execution
handler.

Signed-off-by: Stefano Babic <sbabic@denx.de>