Lua: make logging macro wrappers more printf()-like
Currently, the Lua wrapper methods for calling the macros
wrapping swupdate_notify(), i.e., INFO, ERROR, ..., are limited
to accept a single parameter as message being the string handed
over to swupdate_notify().

In order to have a printf()-like behavior in the Lua realm as
it's available in the C realm, the parameter to those macros has
to be processed by string.format() beforehand, i.e., a formatted
error message from the Lua realm is issued like

    swupdate.error(string.format("message was: %s", msg))

Hence, modify the logging macro Lua wrapper methods so that
string.format() is implicitly called in the C realm prior to
handing over the processed string to the respective logging
macro. This way, formatted messages from the Lua realm can be
issued in a more printf()-like style like

    swupdate.error("message was: %s", msg)

While at it, introduce Lua wrapper methods for the missing WARN
and DEBUG macros.

This change is backwards-compatible and doesn't require any
modification to existing code. The only impact is more
convenience and one call to string.format() per log message
issued from the Lua realm, even if not required.

Signed-off-by: Christian Storm <christian.storm@siemens.com>
Acked-by: Stefano Babic <sbabic@denx.de>