mongoose: set V1 or V2 at runtime
V1 version of REST-API is still more suitable for M2M
updates together with the SWU forwarder handler.
However, it is not currently possible to switch
between them and the API must be selected at compile time,
and this requires different binaries.

Let the user decides which version of API should run
with a command line parameter.

Signed-off-by: Stefano Babic <sbabic@denx.de>
CC: Sami Hartikainen <sami.hartikainen@teleste.com>