UBI handler: support static volumes
UBI volumes can be static or dynamic. Static volumes have a CRC
checksum that is checked when the UBI partition is attached. This gives
additional protection against e.g. accidental overwrites of the MTD
partition.

To support creation of static volumes in SWUpdate, use the handler's
"data" string to indicate whether the volume should be static or
dynamic. If "data" is the literal string "static", a static volume is
created. If it is anything else, a dynamic volume is created.

Note that if the existing volume already has the correct size, nothing
is done. This is normally the right thing to do (the assumption is that
if a static volume is desired, the original image already contains it
as a static volume).

Also update the documentation to reflect this.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Acked-by: Stefano Babic <sbabic@denx.de>