Add a dry-run option to SWUpdate
Add a dry run option to allow to run SWUpdate first without
any change to Hardware. This can be also used during the build
to check the SWU before testing on the target.

During a dry-run mode, all handlers are exchanged with the
dummy handler, that skips the image without doing anything.

Signed-off-by: Stefano Babic <sbabic@denx.de>