hawkbit: allow to set a ciphers suite
As default, curl accepts all implemented ciphers. It is possible
to use a restricted list of ciphers to drop ciphers that are not
considered safe enough. This adds an entry in the configuration file
to set a string with a list of ciphers as they are accepted by
in CURLOPT_SSL_CIPHER_LIST option.

Signed-off-by: Stefano Babic <sbabic@denx.de>