no-downgrading feature
Some projects require that just new software can be installed and
installation of older software is forbidden. For such cases, a new
command line parameter is introduced (-N <current version>) to inform
SWUpdate about the current version.

sw-description already contains a "version" field. This is used
to compare the new and current versions.

Signed-off-by: Stefano Babic <sbabic@denx.de>