Allow to override the dry-run flag by each install
The dryrun flag is set global as command line parameter, but there
are use cases where it is desired to set it for each install request.
In fact, the install can be first checked using the dryrun flag, and if
it was successfully, a normal install is run.

This does not apply when the SWU is local and SWUpdate works in one shot,
because the dryrun flag can be passed as command line parameter and
SWUpdate terminates after its execution.

Signed-off-by: Stefano Babic <sbabic@denx.de>