handlers: handler to update microcontroller via UART
The handler implements a simple protocol to upgrade a microcontroller
via a UART connection. The microcontroller listens to a pin during
reset to check if it should remain in bootloader and waits for
a an upgrade. The handler gets information about the GPIOs for
reset and programming from the sw-description file.

The protocol is described in the header of the source file.

Signed-off-by: Stefano Babic <sbabic@denx.de>