Check if sw-description is really extracted when installed from file
The -c (check) option verifies sw-description and signature. However,
it does not check if the files are correctly extracted before verifying.
If files were already extracted by a previous run, the result can still be
successful even if the SWU is completely wrong.

This does not happen in case of network install because files are cleaned up
after run.

Check that the extraction of sw-description and sw-description.sig is successful
before parsing them.

Signed-off-by: Stefano Babic <sbabic@denx.de>
Reported-by: Karim Lazhard <karim.lazhard@gmail.com>