suricatta: general service (HTTP connector)
This implements a very simple and easy connector to
a backend using HTTP return codes to promote an upgrade.
The standard HTTP codes returned by the server are interpreted
by SWUpdate to check if an update is available.
A generic query to the server is provided to inform the backend
about versions and identification number of the running client
(serial number, MAC address, version running, etc.).

Signed-off-by: Stefano Babic <sbabic@denx.de>