Encryption of ubi volume image
Encryption on ubi volumes is actually broken cause a mismatch between
the size expected to be written and the actual size after decrypting.
This behavior is not related to swupdate: indeed it's kernel api for ubi
volumes that requires to declare the size to be written before to
actually write anything.
Before this patch, the size to be written on disk is retrieved from the
size of the encrypted image when the real size of a decrypted image
is instead smaller.

This patch adds a "decrypted-size" property to the sw-description to
explicitly tell to swupdate to allocate that size for ubi volumes.
This parameter should be updated to the real size of the image just
before assembling the update.

Signed-off-by: Angelo Compagnucci <angelo@amarulasolutions.com>
Acked-by: Stefano Babic <sbabic@denx.de>