channel_curl: make followlink configurable
In most cases, a HTTP redirect should be follow to get
the data to be retrieved. Some protocols require a logic
when a redirect is received, and the automatic logic inside
the libcurl library is not sufficient.

Ann a "nofollow" flag in the channel configuration data (inverted
logic to be compatible with current implementation) to allow to disable
the Curl option.

Signed-off-by: Stefano Babic <sbabic@denx.de>