channel_curl: add redirect as return code and store new URL
A redirect was not verified because the curl library was
set to follow links. After making this configurable
(default=follow links), the returned HTTP code shpuld be managed.
To make prossible to handle the URL in case of
redirect, store the redirect URL into the channel data.

It can be retrieved and handled by the server implementation.

Signed-off-by: Stefano Babic <sbabic@denx.de>