UBI handler: allow mtd device to be specified by name
It is often more convenient to specify the partition by name instead
of by number. This is especially true for boards with multiple chips,
where the numbering may change after an update of the kernel.

If get_mtd_from_device fails, fall back to get_mtd_from_name. In other
words, if mtd partition mtd5 is named "mtd0" (which would be a stupid
thing to do), "mtd0" will still resolve to /dev/mtd0.

Note that currently, the flash handler (i.e., the handler for images
with type = flash) already has this feature. However, for the flash
handler, you have to use 'device' if you specify the mtd number and
'mtdname' if you specify the mtd name. For the ubivol handler (where
the mtd device is specified in the 'partitions' section, not in the
'images' section, so it's not entirely comparable), we choose to
use 'device' both for the mtd number and the mtd name.

The description of this feature is added to sw-description.rst.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Acked-by: Stefano Babic <sbabic@denx.de>