Fix build failure by loading CMS certificate
The correct way to access certificate names is using accessors
provided by openSSL library instead of direct access to fields
inside X509 structure. Changes in openSSL cause build error
because "name" is not anymore part of X509 struct.

Fix it using accessor to get issuer and subject of a
certificate.

Signed-off-by: Stefano Babic <sbabic@denx.de>