hawkBit: add authentication with gateway token
Add support to authenticate to hawkBit using a gateway security token.
Gateway security token is a single token shared between all the devices of a
given tenant. It is useful during development as a simple authentication
method. For more informations see:
https://www.eclipse.org/hawkbit/concepts/authentication/
The gateway token can be passed as a command line parameter with the '-g' or
'--gatewaytoken' options or with the 'gatewaytoken' parameter in the
configuration file.

Signed-off-by: Diego Rondini <diego.rondini@kynetics.com>
Signed-off-by: Marta Todeschini <marta.todeschini@kynetics.com>
Acked-by: Stefano Babic <sbabic@denx.de>