cpio_utils: fix infinite read when interrupting an update
When using the IPC API to do an update from a stream using a compressed
image, a new update cannot be started if the previous one is
interrupted.

The root cause is: When the client closes the socket, the read on the
server side will return 0 and it is treated as a successful read in
gunzip_step(). Thus, it goes into an infinite read loop.

It is fixed by adhering to the contract that an upstream step return
zero in case of EOF.

It can be reproduced using a compressed image:

1. Start the server via "swupdate -e stable,image2".
2. Trigger an update via "swupdate-client image.swu".
3. Cancel the update via Ctrl + C.
4. Trigger another update via "swupdate-client image.swu".

The following is the sw-description file used for the test:

software =
{
    version = "4.0.0";

    hardware-compatibility: ["4.0"];

    stable = {
        image1: {
            images: (
            {
                filename = "rootfs.ext4.gz";
                device = "/dev/mmcblk2p1";
                compressed = true;
                installed-directly = true;
                sha256 = "@SHA256SUM@";
            }
            );
        };
        image2: {
            images: (
            {
                filename = "rootfs.ext4.gz";
                device = "/dev/mmcblk2p2";
                compressed = true;
                installed-directly = true;
                sha256 = "@SHA256SUM@";
            }
            );
        };
    };
}

Signed-off-by: Ben Guan <ben.guan@cn.bosch.com>
Signed-off-by: Mark Jonas <mark.jonas@de.bosch.com>
Acked-by: Stefano Babic <sbabic@denx.de>