parser: function to get a new path of a node
Using links, the path to a node must be computed again. The function
set_find_path() get the string pointed to the link and modify the array
of strings containing the whole path to a node.

A convention is used to traverse the tree upwards. If the string begins
with "..", this means to go up in the tree. Multiple leading ".." are
allowed to go up into the tree.

Signed-off-by: Stefano Babic <sbabic@denx.de>