parser: introduce links in sw-description
There are a lot of use cases when special configurations are needed.
Current sw-description requires that each entry is full described and
does not allow to factorize some parts. This means that in most cases
entries are duplicated multiple times even if changes are minimal
between sections.

This patch introduce links for the root node. If the node is a string
instead of an object, it is interpreted as new link to be follow.
The parser tries to follow the links recursively (a maximum number of 10
links is introduce to avoid deadlocks).

The patch factorizes how to call find_node() for the specific parser.
Instead of using two mechanism (a formatted sting and an array of
strings) for libconfig and json, use an array of strings for both.

Signed-off-by: Stefano Babic <sbabic@denx.de>