handler: bootloader: postpone setting the environment
The bootloader handler sets the environment using an image (ASCII file)
with the list of variables to be set and their values. This can under
some circunstances brick the target because the environment is set in
the middle of un update and not at the end, as it happens when the
variables are listed in the "bootenv" section of sw-description. The
environment is in fact set for the new release and in case of a power
cut or if simply the update is cancelled, the device after a reboot will
try to start with the new environment. In most cases, the new
environment points to the new version of the software that is at least
corrupted orabsent.

This patch changes the behavior and chekcs the values read from the
ASCII file, but not set the environment. It puts the variables in the
global list exactly as the parser does, and the environment will be set
just at the end of the installation.

Signed-off-by: Stefano Babic <sbabic@denx.de>