webserver: introduce a parameter to run postupdate
After a successful update, the Webserver initiates the postupdate
command automatically. There are some use case where this is not desired
because some manual interventions are required on the device before
rebooting. Make this behavior configurable adding a "run-postupdate"
flag in the webserver section of the configuration file (default =
active).

Signed-off-by: Stefano Babic <sbabic@denx.de>