suricatta: general: fix crash without config file
If no configuration data is passed, the server crashes (strdup called
with NULL argument). Without configuration file, no identify data can be
retrieved: the identify data is used to format the query string to the
server, that results Null. This is not correct and the URL is set to the
one passed via command line.

Signed-off-by: Stefano Babic <sbabic@denx.de>