mtd-interface: Properly handle flash devices without lock support
mtd_is_locked() returns a negative value if the flash driver does not support
(or require) locking. In that case, the flash routine should not attempt to
call further (un)lock functions, they will only result in logging like this:

libmtd: error!: MEMUNLOCK ioctl failed for eraseblock 20 (mtd4)
        error 95 (Operation not supported)

To fix the issue, check that the return value is above zero.

Signed-off-by: Mike Looijmans <mike.looijmans@topic.nl>