Drop debian directory as step for inclusion
This is a step for official inclusion in the Debian distribution. Debian
does not require to have package building inside code and the distro
maintains itself the packages. The debian directory is then removed. For
archive purpose it is stored into the -debian branch, but it won't be
needed anymore in future.

Signed-off-by: Stefano Babic <sbabic@denx.de>