Fix setting recovery_status flag in rescue mode
The "recovery_status" flag (bootloader variable) is set when SWUpdate
starts to install images on the target. This works if streaming mode is
not set. When it is set, some images are already installed before
SWUpdate sets the flag. In case of power-on when a single image is used
or in case pf rescue system, the bootloader cannot recognize that an update
was interrupted and tries to start the production software, that can be
corrupted. The rescue is started again just as fallback after failing to
start the production software, but there are cases where this does not
work.

Check if an image is streamed to the target and sets the
"recoverz?status" flag before writing to the storage.

Signed-off-by: Stefano Babic <sbabic@denx.de>