no-reinstalling feature
Some projects require a way to prevent reinstallation of the current firmware
version to prevent repeated updates. For such cases, a new command line
parameter is introduced (-R <current version>) to inform SWUpdate about the
current version. SWUpdate performs a simple string comparision to determine
if the current version is the same as the installed version. We do this so
that no-reinstalling can be used with any versioning format.

We also ensure that the no-downgrading feature doesn't prevent installation
of the existing firmware, it's expected that the no-downgrading feature is
used to enforce a minimum firmware version as opposed to preventing
reinstallation of the existing firmware version, there may be cases where
the minimum version is not the same as the current version, for example a
device may have multiple firmware versions to choose from with the minimum
required version being lower than the current version. For this reason we
enforce the no-downgrading requirement separately from no-reinstalling.

Signed-off-by: James Hilliard <james.hilliard1@gmail.com>