bootloader: support for alternative U-Boot library
Support for U-Boot is done with a libubootenv library generated by the
u-boot-fw-tools package in OE. This library is taken from U-Boot code, but it
must be generaded with the exact U-Boot machine as the bootloader. In
case of mismatch, the environment gets corrupted.

A new library is thought to overcome to these issues and to be more
independent from U-Boot source code. Plan is to have a generic interface
to the environment. Improvements for security are also planned, as
introducing signed environments.

Signed-off-by: Stefano Babic <sbabic@denx.de>