fix the dryrun option for the API for external programs
When using the API for external programs, the network thread
only set the option dryrun in the configuration, so if this
option is set once, it's set forever. Instead we set it each
time, so it's now possible to use the API for external programs
with dryrun then run it withtout dryrun.

Signed-off-by: Philippe Reynes <philippe.reynes@softathome.com>
Acked-by: Stefano Babic <sbabic@denx.de>