Unify bootloader script format
Former script had the syntax:

	<variable> <value>

but it was changed to be confor to u-boot format, also used by U-Boot
tools to generate an environment from a script:

	<variable>=<value>

This is the format used when the environment is updated from handler.
Internally, a script was still generated with the old format and then
passed to the bootloader API. Switch to the new format in all cases and
adjust code in Grub and EBG code.

Signed-off-by: Stefano Babic <sbabic@denx.de>