Increase filename length where it's too small
Most of filename size limitations were unified and unlike
in version 2018.03 we should be able to use files with filenames
longer than 64 bytes, this patch increases the filename size in places
still dependant on the older filename size.

The filename + tmpdir can easily exceed 64 bytes size limit.
Today most of size limitations as in for example install
handlers are unified (for example fully dropped for flash handler: fa6fa9d3),
this patch unifies this limitations even more.

We need this because we usually deal with filenames
with size of around ~100 chars.

Signed-off-by: Krzysztof Michoński <krzysztof.michonski@digitalstrom.com>
Acked-by: Stefano Babic <sbabic@denx.de>