Put the UDS in the directory pointed by TMPDIR
Sockets for IPC can be configured or they are retrieved by systemd, if
active. A compile-time configuration forbids multiple instances of
SWUpdate on the same target because they point to the same sockets.
Change the behavior and creates the sockets in the directory pointed by
the environment variable TMPDIR ("/tmp" is the default). Change is still
compatible with the past, the old names are used if they are set.

Signed-off-by: Stefano Babic <sbabic@denx.de>