bootloader: Return error in case of no valid environment
The swupdate loads environment variables from a persistent storage
(e.g. SPI flash) at first. If there is no environment in the persistent
storage, swupdate try to read environment from a specified file which
contains an initial environment of U-Boot.
If no valid environment can be found from the persistent storage and
given file, swupdate continue to work on an empty environment and store
it to the persistent storage. This leads problem on booting.

This commit fixes the problem by making swupdate return error in case
of no valid environment in persistent storage and given file.

Signed-off-by: Leo Ruan <tingquan.ruan@cn.bosch.com>
Signed-off-by: Mark Jonas <mark.jonas@de.bosch.com>