ubivol_handler: use rsvd_bytes instead of data_bytes
Use vol_info.rsvd_bytes instead of data_bytes. For dynamic volumes
both values are equal, but for static volumes, data_bytes is the
number of stored bytes and rsvd_bytes is the size of the volume.

As a consequence, the check in adjust_volume for whether the volume
size has changed is always true for static volumes, and the volume
gets unecessarily deleted and recreated.

Signed-off-by: Markus Klotzbuecher <markus.klotzbuecher@kistler.com>
Acked-by: Stefano Babic <sbabic@denx.de>