RFC: ubi: add support for atomic volumes rename
This patch introduces a new ubivol image keyword "replaces" that must
contain a UBI volume name. After successfully updating the image to
volume, an atomic swap of the names of 'volume' and 'replaces' is
done. For example

	{
 		filename ="u-boot.img";
                volume ="u-boot_r";
                replaces ="u-boot";
	}

After u-boot.img is successfully installed into the volume
"u-boot_r", the volume 'u-boot_r' is renamed to 'u-boot' and 'u-boot'
is renamed to 'u-boot_r'.

This mechanism allows to implement a double copy update approach
without the need of shared state with the bootloader. For example,
together with U-Boot ubispl load by volume support [1], the SPL can be
configured to always load U-Boot from the volume 'u-boot' without the
need to access the environment and the volume replace functionality
will ensure that this volume name points to the correct image.

[1] https://patchwork.ozlabs.org/patch/1086819/

Signed-off-by: Hamish Guthrie <hamish.guthrie@kistler.com>
Signed-off-by: Markus Klotzbuecher <markus.klotzbuecher@kistler.com>