tools: swupdate-sysrestart to restart system
This tool receives from the progress interface a list of devices that
are updated at the same time to provide a controlled way to restart them
all at once.

The tool reads INFO messages from progress and search for:

	REMOTE:<ip address>

If the pattern is found, the ip address is checked (local addresses are
removed) and added to the list. After a successful update, the tool
initiates a system reboot by connecting to the Webserver of each remote
SWUpdate and sends a reboot command via REST-API. At the end, reboots
itself calling reboot().

Signed-off-by: Stefano Babic <sbabic@denx.de>