Remove the superfluous salt parameter
As it turned out salt is not needed for decryption of encrypted files or
images. Using salt is still recommended for generating keys and IVs from
passwords but was always unneeded for doing actual encryption and
decryption.

Best is to not start with a password at the first place, just random
generate the key and IV like this:

```
key=$(dd if=/dev/random bs=32 count=1 | xxd -ps -cols 64)
iv=$(dd if=/dev/random bs=16 count=1 | xxd -ps -cols 32)
```

Refer to the following discussion for more info:

https://groups.google.com/forum/#!topic/swupdate/JBopZ5LNZis

Signed-off-by: Laszlo Ashin <laszlo@ashin.hu>