rdiff_handler: fix compressed / encrypted delta
When delta file is compressed, the handler does not work because it
checks for the size of the compressed image instead of uncompressed one.
Do not wait for EOF in input buffer, because it cannot be known when the
handler starts: copyfile() simply stops to call the callback when all
data are successful uncompressed.

Signed-off-by: Stefano Babic <sbabic@denx.de>
Acked-by: Christian Storm <christian.storm@siemens.com>