Stream the SWU in a file chosen in sw-description
The -o (--output) parameter allows to store the incoming SWU in a file.
However, if more as one SWU is defined to update different parts of
system (such as system, application, configuration), the file is
overwritten. Let define the output file in sw-description. This requires
that the incoming SWU is first cached into a temporary files just to get
sw-description and after its validation, the output attribute is set.
After that, SWUpdate will redirect the incoming stream to the new
defined file.

Signed-off-by: Stefano Babic <sbabic@denx.de>