Add SSBL Handler
SSBL (Second Stage Boot Loader) is a way outside the bootloader to swap
different software set. A duplicated structure is saved on flash
(currently, just NOR flash is supported) and each of them contains data
for an image to be loaded as first step by a bootloader. The structure
contains an "age" field to check which is the currently active software
set. The handler acts as a switch: it does not install any image and it
registers itself as scripthandler. When called, it checks which is the
inactive SSBL and updates with new data, incrementing the age field to
make it active.

Signed-off-by: Stefano Babic <sbabic@denx.de>