Add Kconfig option for new systemd files
This adds an option for the new systemd service files installation
to menuconfig.

The default configuration for the new SYSTEMD_SYSTEM_UNITDIR setting
is an empty string. This allows a distro buildsystem to provide the
path for systemd system unitdir as an environment variable. An empty
string implies to not install the default systemd configuration files.
The default behavior of make install does not change.