Makefile: rename tools binaries
This adds a "swupdate-" prefix to all binary files in the tools
directory.
Binaries should not be installed as /usr/bin/client or
/usr/bin/progress. On the one hand, it is risky for name conflicts.
On the other hand, the name gives little indication of the meaning of
the binary. This renaming is also required by Debian maintainers.

Another reason for the renaming is that do_install in meta-swupdate is
currently defective. The binaries are installed twice, once with and
once without the swupdate prefix. If the binaries are already installed
correctly named by "make install", the code that renames and duplicates
the binaries can be deleted from meta-swupdate.

This change also fixes an inconsistency in the "make install" target.
The recently added service files call the binaries named swupdate-*.
But these binaries have not yet been installed. The inconsistency was
overlooked because meta-swupdate provides two copies of the binaries.

The swupdate-sysrestart was missing in the .gitignore file.

Signed-off-by: Adrian Freihofer <adrian.freihofer@siemens.com>
Acked-by: Stefano Babic <sbabic@denx.de>