Do not traces messages for progress interface
During downloading a TRACE is sent to the notifier with the goal to
forward it to the progress interface. However, as TRACE, this is also
shown and processed by all notifier causing an overflood of messages on
the console that cna slow down the system.

Move this to a separate notifier and drop it from console notifier.

Signed-off-by: Stefano Babic <sbabic@denx.de>