Revert "systemd generic startup"
Installing the service files from swupdate's Makefile did not proof to
work well. In conjunction with the kbuild-system this approach
introduces some circular dependencies between the menuconfig and the
distro build-system such as Yocto.

Lets keep the systemd service files and the related scripts in the
meta-swupdate repo and add also the new shell configuration features to
meta-swupdate.

This reverts commits
  bdc1d3ca06385a3ccb6e34bc74dc9d431a02d3fa
  75985aab9e0ab7b12185d5d512c68cf74aa0e3b0

Signed-off-by: Adrian Freihofer <adrian.freihofer@siemens.com>