Webserver: monitor and check for lost connection
This solves the following issue:

- Update is initiated by the Webserver
- SWupdate has started and data flows into IPC
- Connection is lost

In this case, the WebBrowser is often restarted by the user via the
refresh button. The old connection is not living anymore, but SWUpdate
does not detect it and waits for data. But when the browser initiates a
new upload, SWUpdate returns with an error due to the running update.

This adds a timer to the webserver connection. As default, connection
monitor is disabled to be compatible with past. Value for the timer can
be passed to the Webserver using the "-t" parameter, as it is already
done for the downloader. Example:

	swupdate -w "-r /www -t 20"

Timeout value can be also passed via the configuration file. The
Webserver sets a timer and checks if some data were transfer since the
last time the timer elapses. If no data is received in an interval equal
to two times the timeout, connection is declared lost and the ipc is
closed. A browser can then initiate a new upload.

Signed-off-by: Stefano Babic <sbabic@denx.de>